fix (solver): fix issue where solver failed if two words found at end of same search
If a prefix such as 'aether' made recursive calls to the 8 grid
positions around it and found 'aethers' (a valid word) at two of these
positions, a copy of the positions seen set was not correctly made,
resulting in the second of these 'aethers' including an extra position.
This led to one too many letters being removed, making it impossible to
complete the solver for the next word in the search.